Move getInstance method to models-utils
