+++ src/ggrc-client/js/components/custom-autocomplete/autocomplete-wrapper.js
@@ -12,7 +12,7 @@ import {
   batchRequests,
 } from '../../plugins/utils/query-api-utils';
 import RefreshQueue from '../../models/refresh_queue';
-import {getInstance} from '../../models/models-extensions';
+import {getInstance} from '../../plugins/utils/models-utils';
 
 export default can.Map.extend({
   currentValue: '',
